refactoring and removal of two unused classes
