--- project-management/src/main/java/life/qbic/projectmanagement/application/SpeciesSearchService.java
@@ -1,10 +1,0 @@
-package life.qbic.projectmanagement.application;
-
-import java.util.List;
-import life.qbic.projectmanagement.domain.model.experiment.vocabulary.Species;
-
-public interface SpeciesSearchService {
-
-    List<Species> fetch(int offset, int limit, String filterText);
-    int getCount(String filterText);
-}
